clean up fixes.test.ts
